Improve error reporting from extension manager
Change the message that is logged when a required extension is not
loaded or found from "Extension not autoloaded: <extension id>" to
"Required extension not found or not loaded: <extension id>".

Change the message that is logged when the extension manager have an
error happening during startup from "Cannot start extension manager:
<error>" to "An error occured during extension manager startup: <error>"

Add the possibility of startExtensionManager returning with an error
if the Dispatcher failed to properly add the ExtensionManagerWatcher
or ExtensionManagerRunner.

Fallback to logging the startExtensionManager error message with a
serverity of INFO only if extensions are required to be disabled.

Issue osquery/osquery#5679