Change the retry system for expectTrue, try_open and open
The timeout concept was confusing, because it wasn't really
waiting for that amount of time and then interrupting the
tentatives. It was just trying timeout/interval times
which by default, in the expectTrue case, was 800 times, which
is a lot and compounds to a lot if the function tested did some
tentatives too.

timeout is now called attemps, and the default values have been
changed to something a bit saner.