CMake: select most recent package available
By default CMake find_package do not make any attempt
to sort the package versions it finds and it takes the first.

NATURAL order will sort by version as if it was a whole number,
without the dots in between.
We also explicitly set that the sort direction is descending,
so that the newest version is actually selected.